Throw an error during validation when there is no certificate
When there is no certificate (ex: if the p12 file is missing one), the `certificates` array is empty. This causes an exception later in the code along the lines of "TypeError: Cannot read property 'key' of undefined" because `certificate` is undefined. This commit specifically checks for a missing certificate and throws a clearer error message instead.

Tested by adding a new unit test.